+++ lib/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "helium-console"
+version = "0.1.5"
+authors = ["Louis Thiery <louis@helium.com>"]
+edition = "2018"
+
+[lib]
+path = "lib.rs"
+
+[dependencies]
+structopt = { version = "0.3.2", default-features = false }
+reqwest = { version = "0.10", features = ["json"] }
+tokio = { version = "0.2", features = ["full"] }
+serde =  "1"
+serde_derive = "1"
+serde_json = "1"
+hex = "0.4"
+base64 = "0.11"
+uuid = "0.8"
+toml = "0.5"
+oauth2 = "2.0"
+rand = "0.7.3"
+url = "^1.7.2"
+hyper = "0.13.3"
+prettytable-rs = "^0.8"
